pyproject.toml
[tool.poetry]
name = "discord-musicbot"
version = "0.1.0"
description = ""
authors = ["jinkjonks <jennyli12320@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
